fixed by hikariatama

+++ abstract.py
@@ -1,53 +1,60 @@
-#     ▄▀█ █▀▄▀█ █▀█ █▀█ █▀▀
-#     █▀█ █░▀░█ █▄█ █▀▄ ██▄
-#          
-#             © Copyright 2022
+# ▄▀█ █▀▄▀█ █▀█ █▀█ █▀▀
+# █▀█ █░▀░█ █▄█ █▀▄ ██▄
 #
-#          https://t.me/the_farkhodov 
+#             © Copyright 2022
 #
-#🔒 Licensed under the GNU GPLv3
+#          https://t.me/the_farkhodov
+#
+# 🔒 Licensed under the GNU GPLv3
 # 🌐 https://www.gnu.org/licenses/agpl-3.0.html
 
-
+# scope: inline
 # scope: hikka_only
-# scope: hikka_min 1.3.1
+# scope: hikka_min 1.3.0
 # meta developer: @amoremods
 # meta banner: https://imgur.com/EzGrs6h
 
-
 from telethon import events
 from .. import utils, loader
-import re, asyncio, os
 
-cute = 5475546549
-meta = " Made by @amoremods❤"
+chat = "amoreconverterbot"
 
 
-@loader.tds
 class AbstractMod(loader.Module):
-	"""Напишет красивый конспект на листке"""
-	
-	strings = {"name": "Конспект",
-                   "developer": "@sngscamer",
-                       }
-	
-	async def client_ready(self, client, db):
-		self.db = db
-	
-	@loader.owner
-	async def konspcmd(self, message):
-		""" .konsp text"""
+    """Write a beautiful summary on a notebook"""
 
-		text = utils.get_args_raw(message) 
-		async with message.client.conversation(cute) as conv:
-			await utils.answer(message, '✍ Пишу конспекст...')
-			response = conv.wait_event(events.NewMessage(incoming=True, from_users=cute, chats=cute))
-			amore = await message.client.send_message(cute, "📖 Konspekt yozish")
-			await utils.answer(message, 'Секунду...')
-			amore = await message.client.send_message(cute, text)
-			response = await response
-			await message.client.send_file(message.to_id, response.media, caption=meta)
-			await response.delete()
-			amore.delete()
-			await message.delete()
-			await self.client.delete_dialog(cute)
+    strings = {
+        "name": "Abstract",
+        "processing": (
+            "<emoji document_id='6318766236746384900'>🕔</emoji> <b>Processing...</b>"
+        ),
+    }
+
+    @loader.owner
+    @loader.command(ru_doc="<текст> - Создать конспект")
+    async def konspcmd(self, message):
+        """<text> - Create summary"""
+        text = utils.get_args_raw(message)
+        message = await utils.answer(message, self.strings("processing"))
+        async with self._client.conversation(chat) as conv:
+            msgs = []
+            msgs += [await conv.send_message("/start")]
+            msgs += [await conv.get_response()]
+            msgs += [await conv.send_message("📖 Konspekt yozish")]
+            msgs += [await conv.send_message(text)]
+            m = await conv.get_response()
+
+        await self._client.send_file(
+            message.peer_id,
+            m.media,
+            caption="❤ @amoremods",
+            reply_to=message.reply_to_msg_id,
+        )
+
+        for msg in msgs + [m]:
+            await msg.delete()
+
+        if message.out:
+            await message.delete()
+
+        await self.client.delete_dialog(chat)
